+++ src/strands/experimental/bidi/models/gemini_live.py
@@ -40,7 +40,7 @@ from ..types.events import (
     BidiUsageEvent,
     ModalityUsage,
 )
-from .bidi_model import BidiModel, BidiModelTimeoutError
+from .model import BidiModel, BidiModelTimeoutError
 
 logger = logging.getLogger(__name__)
