+++ tests/strands/experimental/bidi/models/test_gemini_live.py
@@ -13,7 +13,7 @@ import unittest.mock
 import pytest
 from google.genai import types as genai_types
 
-from strands.experimental.bidi.models.bidi_model import BidiModelTimeoutError
+from strands.experimental.bidi.models.model import BidiModelTimeoutError
 from strands.experimental.bidi.models.gemini_live import BidiGeminiLiveModel
 from strands.experimental.bidi.types.events import (
     BidiAudioInputEvent,
